+++ app/src/features/cards/Card.module.scss
@@ -3,6 +3,10 @@
 }
 
 .ChartCard {
-  padding: 20px;
   margin: 10px;
+}
+
+.Breadcrumbs {
+  padding-bottom: 6px !important;
+  padding-top: 6px !important;
 }
