+++ app/src/features/cards/DisparityBarChartCard.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { VariableId } from "../../utils/variableProviders";
-import TwoVarBarChart from "../charts/TwoVarBarChart";
+import DisparityBarChart from "../charts/DisparityBarChart";
 import Card from "@material-ui/core/Card";
 import Typography from "@material-ui/core/Typography";
 import styles from "./Card.module.scss";
@@ -11,6 +11,7 @@ import {
   DATASET_PRE_FILTERS,
   DATA_CATALOG_PAGE_LINK,
 } from "../../utils/urlutils";
+import { CardContent } from "@material-ui/core";
 
 function DisparityBarChartCard(props: {
   dataset?: Row[];
@@ -22,35 +23,40 @@ function DisparityBarChartCard(props: {
 }) {
   return (
     <Card raised={true} className={styles.ChartCard}>
-      <Typography gutterBottom className={styles.CardHeader}>
-        Disparities in {props.variableDisplayName} for{" "}
-        <b>{props.breakdownVarDisplayName}</b>
-      </Typography>
-      {!props.dataset && (
-        <Alert severity="warning">
-          Missing data means that we don't know the full story.
-        </Alert>
-      )}
-      {props.dataset && (
-        <>
-          <TwoVarBarChart
-            data={props.dataset}
-            thickMeasure="population_pct"
-            thinMeasure={props.variableId}
-            thickMeasureDisplayName="Population %"
-            thinMeasureDisplayName={props.variableDisplayName + " as % of Geo"}
-            breakdownVar={props.breakdownVar}
-            breakdownVarDisplayName={props.breakdownVarDisplayName}
-          />
-          <LinkWithStickyParams
-            to={`${DATA_CATALOG_PAGE_LINK}?${DATASET_PRE_FILTERS}=${props.datasetIds.join(
-              ","
-            )}`}
-          >
-            View Data Sources
-          </LinkWithStickyParams>
-        </>
-      )}
+      <CardContent>
+        <Typography gutterBottom className={styles.CardHeader}>
+          Disparities in {props.variableDisplayName} for{" "}
+          <b>{props.breakdownVarDisplayName}</b>
+        </Typography>
+        {!props.dataset && (
+          <Alert severity="warning">
+            Missing data means that we don't know the full story.
+          </Alert>
+        )}
+        {props.dataset && (
+          <>
+            <DisparityBarChart
+              data={props.dataset}
+              thickMeasure="population_pct"
+              thinMeasure={props.variableId}
+              thickMeasureDisplayName="Population %"
+              thinMeasureDisplayName={
+                props.variableDisplayName + " as % of Geo"
+              }
+              breakdownVar={props.breakdownVar}
+              breakdownVarDisplayName={props.breakdownVarDisplayName}
+            />
+            <LinkWithStickyParams
+              target="_blank"
+              to={`${DATA_CATALOG_PAGE_LINK}?${DATASET_PRE_FILTERS}=${props.datasetIds.join(
+                ","
+              )}`}
+            >
+              View Data Sources
+            </LinkWithStickyParams>
+          </>
+        )}
+      </CardContent>
     </Card>
   );
 }
