+++ app/src/features/reports/VariableReport.tsx
@@ -9,7 +9,7 @@ import { DropdownVarId } from "../../utils/MadLibs";
 import { Breakdowns } from "../../utils/Breakdowns";
 import VariableProvider from "../../utils/variables/VariableProvider";
 import { USA_FIPS, Fips } from "../../utils/Fips";
-import MapNavChart from "../charts/MapNavChart";
+import MapNavCard from "../cards/MapNavCard";
 import Alert from "@material-ui/lab/Alert";
 import Card from "@material-ui/core/Card";
 
@@ -87,21 +87,16 @@ function VarGeoReport(props: {
                   md={props.vertical ? 12 : 6}
                   className={styles.PaddedGrid}
                 >
-                  <Card
-                    raised={true}
-                    style={{ margin: "10px", padding: "20px" }}
-                  >
-                    <MapNavChart
-                      data={dataset}
-                      varField={variableId}
-                      varFieldDisplayName={variableDisplayName}
-                      fips={fips}
-                      updateFipsCallback={(fips: Fips) => {
-                        setFips(fips);
-                        props.updateStateCallback(fips.getStateFipsCode());
-                      }}
-                    />
-                  </Card>
+                  <MapNavCard
+                    data={dataset}
+                    varField={variableId}
+                    varFieldDisplayName={variableDisplayName}
+                    fips={fips}
+                    updateFipsCallback={(fips: Fips) => {
+                      setFips(fips);
+                      props.updateStateCallback(fips.getStateFipsCode());
+                    }}
+                  />
                 </Grid>
                 <Grid
                   item
